+++ src/store/slidersStore.js
@@ -32,10 +32,20 @@ const useSlidersStore = create((set, get) => ({
   addSlide: async data => {
     if (isDataValid(data)) {
       try {
+        set(() => {
+          return {
+            loading: true,
+          };
+        });
         const response = await axios.post('/slider_main', data, {
           headers: {
             'Content-Type': 'multipart/form-data',
           },
+        });
+        set(() => {
+          return {
+            loading: false,
+          };
         });
         return response;
       } catch (error) {
@@ -47,10 +57,20 @@ const useSlidersStore = create((set, get) => ({
   editSlide: async (id, data) => {
     if (isDataValid(data)) {
       try {
+        set(() => {
+          return {
+            loading: true,
+          };
+        });
         const response = await axios.put(`/slider_main/${id}`, data, {
           headers: {
             'Content-Type': 'multipart/form-data',
           },
+        });
+        set(() => {
+          return {
+            loading: false,
+          };
         });
         return response;
       } catch (error) {
@@ -60,13 +80,29 @@ const useSlidersStore = create((set, get) => ({
   },
 
   deleteSlide: async id => {
-    const response = await axios.delete(`/slider_main/${id}`);
-    set(() => {
-      return {
-        slides: get().slides.filter(slide => slide.id !== id),
-      };
-    });
-    return response;
+    if (id) {
+      try {
+        set(() => {
+          return {
+            loading: true,
+          };
+        });
+        const response = await axios.delete(`/slider_main/${id}`);
+        set(() => {
+          return {
+            slides: get().slides.filter(slide => slide.id !== id),
+          };
+        });
+        set(() => {
+          return {
+            loading: false,
+          };
+        });
+        return response;
+      } catch (error) {
+        throw new Error(error);
+      }
+    }
   },
 }));
